% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_SpectralTreatments.R
\name{apply_SpectralTreatments}
\alias{apply_SpectralTreatments}
\title{Filtering methods for eSpectra objects}
\usage{
apply_SpectralTreatments(eSpectra_A, eSpectra_B)
}
\arguments{
\item{eSpectra_A}{Object of class \code{\link{eSpectra}}.}

\item{eSpectra_B}{}
}
\description{
#' (NOT FULLY IMPLEMENTED) 
 
This method applies spectral pretreatments from a class \code{\link{eSpectra}} object 'A' to another \code{\link{eSpectra}} object 'B'
}
\examples{
\dontrun{
data("eSpectraExample")
filteredSpectra <- filter_eSpectra(eSpectraExample,c('SNV','C-hull'))
result <- apply_SpectralTreatments(eSpectraExample,filteredSpectra)

result
}
}
\author{
Mario Fajardo.
}
