% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bag_eSpectra.R
\name{bag_eSpectra}
\alias{bag_eSpectra}
\title{Split eSpectra object into a bootstrapped sample}
\usage{
bag_eSpectra(eSpectra = NULL, val_split = NULL, mode = "random",
  iters = 50, lhcIter = 15000, ...)
}
\arguments{
\item{eSpectra}{A \code{\link{eSpectra}} object.}

\item{val_split}{percentage of the dataset that will be used for validation}

\item{mode}{currently two methods implemented, 'lhc' for latin hypercube sampling and 'random' for random sampling (default). The latin hypercube sample design uses \code{\link{clhs}} function and is applied to the first 5 principal component scores of the eSpectra Spectral information}

\item{lhcIter}{Latin hypercube iterations (if mode = 'lhc').For testing purposes is recommended to lower this value.}
}
\value{
A list containing:
\item{Train}{Training IDs}
\item{Validation}{Validation IDs}
\item{BootSamples}{Bootstrapped samples if validation='external'}
}
\description{
Split eSpectra object into a bootstrapped sample
}
\examples{
\dontrun{
data("eSpectraExample")

Bag_Sample <- bag_eSpectra(eSpectra=eSpectraExample,val_split = 0.25,mode = 'random',iters = 10)

Bag_Sample

}

}
\author{
Mario Fajardo
}
