% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_models.R
\name{boot_models}
\alias{boot_models}
\title{Train Bootstrapped models}
\usage{
boot_models(eSpectra = eSpectra, bootSample = NULL, target = NULL,
  ...)
}
\arguments{
\item{eSpectra}{A \code{\link{eSpectra}} object.}

\item{bootSample}{a bootsample obtained with \code{\link{bag_eSpectra}}}

\item{target}{character vector of variable to be predicted (must be in eSpectra@Properties)}
}
\value{
A list containing:
\item{predicted}{predictions}
\item{observed}{observations}
\item{acc}{accuracy indices}
}
\description{
Train Bootstrapped models
}
\examples{
\dontrun{
data("Almonds")

Bag_Sample <- bag_eSpectra(Almonds,val_split = 0.25,mode = 'random',iters = 10)

Models <- boot_models(eSpectra = Almonds,bootSample = Bag_Sample,target = 'OilContent')

Models
}

}
\author{
Mario Fajardo
}
