% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compSpec.R
\name{compSpec}
\alias{compSpec}
\title{Reduce column dimensions}
\usage{
compSpec(spectra, window)
}
\arguments{
\item{spectra}{matrix where each row is a spectrum and each column a wavelength}

\item{window}{window size over which the spectra will be averaged}
}
\description{
This function reduces the column dimensions
}
\author{
Brendan Malone
}

