% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSpectra.R
\docType{class}
\name{eSpectra}
\alias{eSpectra}
\title{Formal S4 class eSpectra}
\description{
Formal S4 class eSpectra
}
\section{Slots}{

\describe{
\item{\code{Meta}}{A character vector with dataset details}

\item{\code{Instrument}}{A character vector specifying de name of the Instrument where the spectram was obtained}

\item{\code{Spectra}}{A numeric matrix which contains absorbance/reflectance soil information}

\item{\code{Bands}}{A character vector specifying Band names for the spectra}

\item{\code{Units}}{A character vector with the units of each Band.}

\item{\code{Id}}{A character vector specyfing each spectrum Id.}

\item{\code{RowsAreSpectra}}{Logical}

\item{\code{Type}}{character vector specifying if the information is expressed in Absorbance, Reflectance, Energy or others (Normalized, EPO, etc.) .}

\item{\code{Treatment}}{Spectral Treatments that had being applied to the spectra}

\item{\code{Properties}}{data.frame with Properties associated with each spectrum}
}}

\author{
Mario Fajardo.
}
