% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter_spectra.R
\name{filter_spectra}
\alias{filter_spectra}
\title{Filtering functions for spectra}
\usage{
filter_spectra(spectra, type, n = 11, p = 2, m = 0, res = NULL,
  specType = NULL)
}
\arguments{
\item{spectra}{matrix where each row is a spectrum and each column is a separate wavelength. It is important that these wavelengths be at least very nearly evenly spaced.}

\item{If}{type = 'S-Golay' then n,p,m parameters control \code{\link{sgolayfilt}} window size, polynomial order and derivative order respectively.}

\item{If}{type = 'Wavelet' res level to be extracted from wavelet decomposition model, see \code{\link{accessC.wd}}}
}
\description{
This function calls different filters on matrices in which each row is a separate spectrum
depending on the value of 'type':
\itemize{
\item 'S-Golay' for Sativsky Golay filter using \code{\link{sgolayfilt}},
\item 'Wavelet' for Wavelet decomposition using \code{\link{accessC.wd}},
\item 'MSC' for Multiplicative scatter correction,
\item 'SNV' for Standard Normal Variate transformation.
\item 'C-hull' for Continuum removal by removing the convex hull
}
}
\author{
Brendan Malone, Budiman Minasny, Michael Nelson, Sebastian Campbell and Mario Fajardo.
}
\references{
Savitzky, Abraham and Golay, M. J. E. (1964) Smoothing and Differentiation of Data by Simplified Least Squares Procedures, Analytical Chemistry, 36, pp. 1627-1629.
}

