% Generated by roxygen2 (4.0.2): do not edit by hand
\name{mscBLC}
\alias{mscBLC}
\title{Multiplicative scatter correction (MSC)}
\usage{
mscBLC(spectra)
}
\arguments{
\item{spectra}{a matrix or data.frame with wavelengths as columns and spectra as rows}
}
\description{
Multiplicative scatter correction (MSC)
}

