% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,SoilSpectra-method}
\alias{plot,SoilSpectra-method}
\title{plot method for SoilSpectra objects}
\usage{
\S4method{plot}{SoilSpectra}(x, y, ...)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}
}
\description{
This method generates a plot for a Spectra Object
}
\examples{
 
\dontrun{
data("SoilSpectraExample")
plot(SoilSpectraExample)
plot(SoilSpectraExample,type='l')
}
}
\author{
Mario Fajardo.
}
