% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPredicions.R
\name{plotPredicions}
\alias{plotPredicions}
\title{Plot bootstrapped model predictions}
\usage{
plotPredicions(models, formula = FALSE, formulaCoords = NULL)
}
\arguments{
\item{models}{list with three elements (predicted, observed and accuracy data) obtained after using \code{\link{boot_models}} function.}

\item{formula}{Should a formula with a linear regression be presented (Logic)}

\item{formulaCoords}{x and y coordinates of Formula (if formula=TRUE)}
}
\value{
A \code{\link{ggplot}} object
}
\description{
Plot bootstrapped model predictions
}
\author{
Mario Fajardo
}
