% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\docType{methods}
\name{points,eSpectra-method}
\alias{points,eSpectra-method}
\title{points method for eSpectra objects}
\usage{
\S4method{points}{eSpectra}(x, ...)
}
\arguments{
\item{eSpectra}{Object of class \code{\link{eSpectra}}.}

\item{ID}{numeric or character vector with ID of spectra to plot (default=1).}
}
\description{
This method generates a plot.eSpectra for a Spectra Object
}
\examples{
 
\dontrun{
data("eSpectraExample")
plot(eSpectraExample)
points(eSpectraExample)
}
}
\author{
Mario Fajardo.
}
