% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,SoilSpectra-method}
\alias{print,SoilSpectra-method}
\title{print method for SoilSpectra objects}
\usage{
\S4method{print}{SoilSpectra}(x)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}
}
\description{
This method generates a description for a Spectra Object
}
\examples{
 
\dontrun{
data("SoilSpectraExample")
print(SoilSpectraExample)
}
}
\author{
Mario Fajardo.
}
