% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\docType{methods}
\name{print,eSpectra-method}
\alias{print,eSpectra-method}
\title{print method for eSpectra objects}
\usage{
\S4method{print}{eSpectra}(x)
}
\arguments{
\item{eSpectra}{Object of class \code{\link{eSpectra}}.}
}
\description{
This method generates a description for a Spectra Object
}
\examples{
 
\dontrun{
data("eSpectraExample")
print(eSpectraExample)
}
}
\author{
Mario Fajardo.
}
