% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,SoilSpectra-method}
\alias{show,SoilSpectra-method}
\title{show method for SoilSpectra objects}
\usage{
\S4method{show}{SoilSpectra}(object)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}
}
\description{
This method generates a description for a Spectra Object
}
\author{
Mario Fajardo.
}
