% Generated by roxygen2 (4.0.2): do not edit by hand
\name{snvBLC}
\alias{snvBLC}
\title{Standard Normal Variate (SNV) transformation}
\usage{
snvBLC(spectra)
}
\arguments{
\item{spectra}{a matrix or data.frame with wavelengths as columns and spectra as rows}
}
\description{
Standard Normal Variate (SNV) transformation
}

