% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoilSpectra.R
\docType{class}
\name{SoilSpectra}
\alias{SoilSpectra}
\title{Formal S4 class SoilSpectra}
\description{
Formal S4 class SoilSpectra
}
\section{Slots}{

\describe{
\item{\code{Instrument}}{A character vector specifying de name of the Instrument where the spectram was obtained}

\item{\code{Spectra}}{A numeric matrix which contains absorbance/reflectance soil information}

\item{\code{Wavelength}}{A character vector specifying wavlength names}

\item{\code{Range}}{A character vector section of the electromagnetic spectrum used e.g., Vis-NIR, NIR, MIR, etc.}

\item{\code{Wavenumber}}{A character vector specifying Wavenumber names}

\item{\code{Id}}{A character vector specyfing each spectrum Id.}

\item{\code{RowsAreSpectra}}{Logical}

\item{\code{Type}}{character vector specifying if the information is expressed in Absorbance, Reflectance or Modified (Normalized, EPO, etc.) .}

\item{\code{Treatment}}{Spectral Treatments that had being applied to the spectra}
}}

\author{
Mario Fajardo, Brendan Malone, Budiman Minasny and Edward Jones.
}
