% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra2colour.R
\name{spectra2colour}
\alias{spectra2colour}
\title{Method that extracts RGB and Munsell colour from \code{\link{eSpectra}} objects.}
\usage{
spectra2colour(eSpectra, ...)
}
\arguments{
\item{eSpectra}{A \code{\link{eSpectra}} object.}
}
\description{
Converts spectra reflectance into RGB and Munsell colours
}
\examples{
\dontrun{
data("eSpectraExample")

spectra2colour(eSpectraExample)

spectra2colour(eSpectraExample[23]) 
}
}
\author{
Michael Nelson and Mario Fajardo
}
