% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\docType{methods}
\name{[,eSpectra,ANY,ANY-method}
\alias{[,eSpectra,ANY,ANY-method}
\title{eSpectra subset method}
\usage{
\S4method{[}{eSpectra,ANY,ANY}(x, i, j, ..., drop = T)
}
\description{
eSpectra subset method
}
