% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{tidy}
\alias{tidy}
\title{tidy method for eSpectra objects}
\usage{
tidy(eSpectra)
}
\arguments{
\item{eSpectra}{A \code{\link{eSpectra}} object.}
}
\value{
A ready to use longFormat data.frame suited for further manipulation by packages like dplyr, reshape2, ggplot2 and ggvis
}
\description{
tidy method for eSpectra objects
}
\examples{
\dontrun{
require(ggplot2)
data("eSpectraExample")

TidySpec <- tidy(eSpectraExample)

head(TidySpec)

ggplot(TidySpec)+geom_line(aes(x = nm, y = Reflectance, group = ID,colour=clay))

}





}
\author{
Mario Fajardo
}
