% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wavelet_smooth.R
\name{wavelet_smooth}
\alias{wavelet_smooth}
\title{Wavelet smoothing}
\usage{
wavelet_smooth(spectra, res)
}
\arguments{
\item{spectra}{a matrix or data.frame where the columns are wavelengths and the rows, individual spectra}

\item{res}{level (an integer >1) to be extracted from wavelet decomposition model}
}
\description{
Wavelet smoothing
}
\note{
This function will be deprecated in the next version, please use \code{\link{filter_spectra}} instead
}
\author{
Brendan Malone
}

