% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_SpectralTreatments.R
\name{apply_SpectralTreatments}
\alias{apply_SpectralTreatments}
\title{Filtering methods for SoilSpectra objects}
\usage{
apply_SpectralTreatments(SoilSpectra_A, SoilSpectra_B)
}
\arguments{
\item{SoilSpectra_A}{Object of class \code{\link{SoilSpectra}}.}

\item{SoilSpectra_B}{}
}
\description{
This method applies spectral pretreatments from a class \code{\link{SoilSpectra}} object 'A' to another \code{\link{SoilSpectra}} object 'B'
}
\examples{
\dontrun{
data("SoilSpectraExample")
filteredSpectra <- filter_SoilSpectra(SoilSpectraExample,c('SNV','C-hull'))
result <- apply_SpectralTreatments(SoilSpectraExample,filteredSpectra)

result
}
}
\author{
Mario Fajardo.
}
