% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaUnderCurve.R
\name{AreaUnderCurve}
\alias{AreaUnderCurve}
\title{Spectra normalisation via continuum estimation}
\usage{
AreaUnderCurve(SoilSpectra, AUC = TRUE, ...)
}
\arguments{
\item{AUC}{A logical that is expressed as the Area Under the Curve (AUC) or Over the Curve is desired, default is set to TRUE i.e., AUC}

\item{Spectra}{A \code{\link{SoilSpectra}} object.}
}
\value{
Returns a list with 8 elements:
\itemize{
\item{hull}{ A \code{\link{SoilSpectra}} object with hull spectra (equivalent to \code{\link{filterSpectra}} output)}
\item{raw}{ A \code{\link{SoilSpectra}} object with the input spectra}
\item{polygons}{ Vertices for constructing the polygons of each normalised spectral feature}
\item{contSpectra} {A \code{\link{SoilSpectra}} object with the Continuum removed spectra}
\item{areas}{ areas of the polygons}
\item{depths}{ difference between the highest reflectance/absorbance and the lower of each polygon}
\item{slopes}{ slopes between the highest reflectance/absorbance and the lower of each polygon}

}
}
\description{
Generic Function for:
\itemize{
\item 1. Fitting continuum from spectral hull features
\item 2. Outputting continuum removed or normalised spectra
\item 3. Estimating normalised hull feature polygon vertices
}
}
\details{
This function is useful in studies for example where spectral features of phenomena need to be investigated
Selecting the convex hull features in the right order has been an ongoing teething issue.
}
\examples{
\dontrun{
data("SoilSpectraExample")
#Select a specific area
SelectedArea <- SoilSpectraExample[122:123,1950:2220]


#Run the continuum fitting function for Area Under the Curve
spec.CR<- AreaUnderCurve(SelectedArea,AUC=T)


plot(SoilSpectraExample[122:123,])
points(spec.CR$raw[1],col='blue',type='l',lwd=2)
points(spec.CR$raw[2],col='red',type='l',lwd=2)


points(spec.CR$continuum[1],col='blue',type='l')
points(spec.CR$continuum[2],col='red',type='l')



#check Differences
plot(spec.CR$hull)
polygon(spec.CR$polygons$`122`,col='blue')
polygon(spec.CR$polygons$`123`,col='red')

spec.CR$areas
spec.CR$slopes
spec.CR$depths





#Run the continuum fitting function for Area Over the curve
SelectedArea <- SoilSpectraExample[122:123,2180:2230]

spec.CR<- AreaUnderCurve(SelectedArea,AUC=F)


plot(SoilSpectraExample[122:123,])
points(spec.CR$raw[1],col='blue',type='l',lwd=2)
points(spec.CR$raw[2],col='red',type='l',lwd=2)


points(spec.CR$continuum[1],col='blue',type='l')
points(spec.CR$continuum[2],col='red',type='l')



#check Differences
plot(spec.CR$hull)
polygon(spec.CR$polygons$`122`,col='blue')
polygon(spec.CR$polygons$`123`,col='red')

spec.CR$areas
spec.CR$slopes
spec.CR$depths
}

}
\author{
Mario Fajardo and Brendan Malone
}
