% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaUnderCurve.R
\name{AreaUnderCurve}
\alias{AreaUnderCurve}
\title{Spectra normalisation via continuum estimation}
\usage{
AreaUnderCurve(SoilSpectra, AUC = TRUE, ...)
}
\arguments{
\item{AUC}{A logical that is expressed as the Area Under the Curve (AUC) or Over the Curve is desired, default is set to TRUE i.e., AUC}

\item{Spectra}{A \code{\link{SoilSpectra}} object.}
}
\value{
Returns a list with 8 elements:
\itemize{
\item{ContRemoved}{The normalised or continuum removed spectra}
\item{InputSpec}{The input spectra}
\item{continuum}{The fitted continuum}
\item{polygon}{Vertices for constructing the polygon of the normalised spectral feature}
\item{area}{area of the polygon}
\item{depth}{difference between the highest reflectance/absorbance and the lower}
\item{slope}{slope between the highest reflectance/absorbance and the lower}

}
}
\description{
Generic Function for:
\itemize{
\item 1. Fitting continuum from spectral hull features
\item 2. Outputting continuum removed or normalised spectra
\item 3. Estimating normalised hull feature polygon vertices
}
}
\details{
This function is useful in studies for example where spectral features of phenomena need to be investigated
Selecting the convex hull features in the right order has been an ongoing teething issue.
}
\examples{
\dontrun{
data("SoilSpectraExample")
#Select a specific area 
SelectedArea <- SoilSpectraExample[123,2079:2267]

plot(SelectedArea)

#Run the continuum fitting function
spec.CR<- AreaUnderCurve(SelectedArea,AUC=T)
# Draw the continuum
lines(SelectedArea@Wavelength,spec.CR$`123`$continuum,col="red")

#Plot continuum removed spectrum
plot(SelectedArea@Wavelength,spec.CR$`123`$c.hull,type="l", xlab="wavelength", ylab="reflectance")
polygon(spec.CR$`123`$polygon,col='red')
}
}
\author{
Mario Fajardo and Brendan Malone
}
