% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chBLC.R
\name{chBLC}
\alias{chBLC}
\title{Continuum removal by removing the convex hull}
\usage{
chBLC(spectra)
}
\arguments{
\item{spectra}{a matrix or data.frame with wavelengths as columns and spectra as rows}
}
\description{
This function applies a linearly interpolated convex hull to the spectra and returns the ratio of the deviation to the hull value
}
\note{
This function will be deprecated in the next version, please use \code{\link{filter_spectra}} instead
}

