% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chBLC_ext.R
\name{chBLC_ext}
\alias{chBLC_ext}
\title{Spectra normalisation via continuum estimation}
\usage{
chBLC_ext(spectra, lower = 350, upper = 2500, AUC = TRUE)
}
\arguments{
\item{spectra}{A single spectrum of class \code{data.frame}. Column values represent spectral values per wavelength unit.}

\item{lower}{numeric; The lower wavelength bound of the spectral feature or region of interest.}

\item{upper}{numeric; The upper wavelength bound of the spectral feature or region of interest.}
}
\value{
Returns a list with 8 elements:
\itemize{
\item{wave}{ The sequence of spectral wavelengths of the region of interest.}
\item{c.hull}{ The normalised or continuum removed spectrum}
\item{raw.spec}{ The raw spectrum}
\item{continuum}{ The fitted continuum}
\item{polygon}{ Vertices for constructing the polygon of the normalised spectral feature}
\item{area}{ NEEDS A GOOD DESCRIPTION AND REFERENCES HOPEFULLY}
\item{depth}{ NEEDS A GOOD DESCRIPTION AND REFERENCES HOPEFULLY}
\item{slope}{ NEEDS A GOOD DESCRIPTION AND REFERENCES HOPEFULLY}

}
}
\description{
Generic Function for:
\itemize{
\item 1. Fitting continuum from spectral hull features
\item 2. Outputting continuum removed or normalised spectra
\item 3. Estimating normalised hull feature polygon vertices
}
}
\details{
This function is useful in studies for example where spectral features of phenomena need to be investigated
Selecting the convex hull features in the right order has been an ongoing teething issue.
}
\note{
{This function works similarly to the other functions in this package where a continuum is fitted to data from which raw spectra values are removed from it. However, this function provides extra useful output. It is currently only scripted to run on a single spectrum.}
}
\examples{
data(soilSpecDat)

#Curation
nc <- ncol(soilSpecDat)
spectra <- soilSpecDat[, 5:nc] #just get the spectra values
wavelength <- seq(350, 2500, by = 1) #wavelength sequence
colnames(spectra) <- wavelength #append column names
#Trim spectra table to a specified region of interest
lower<- 2079
upper<- 2267
specTrim<- trimSpec(spectra, wavlimits=range(lower:upper)) 
plot(seq(from=lower, to=upper, by=1),specTrim[1,],type="l") #FIrst Spectrum

#Run the continuum fitting function
spec.CR<- chBLC_ext(specTrim[1,], lower, upper) #First spectrum
# Draw the continuum
lines(seq(from=lower, to=upper, by=1),spec.CR$continuum,col="red") 

#Plot continuum removed spectrum
plot(seq(from=lower, to=upper, by=1),spec.CR$c.hull,type="l", xlab="wavelength", ylab="reflectance")
polygon(spec.CR$polygon,col='red')
}
\author{
Mario Fajardo and Brendan Malone
}
\references{
{
More information and examples of this function can be found in the Soil Spectral Inference training materials and references therein from the Sydney Soil Schools.
}
}

