% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compSpec.R
\name{compSpec}
\alias{compSpec}
\title{Reduce column dimensions}
\usage{
compSpec(spectra, window)
}
\arguments{
\item{spectra}{matrix where each row is a spectrum and each column a wavelength}

\item{window}{window size over which the spectra will be averaged}
}
\description{
This function reduces the column dimensions by averaging values inside a window of a specified size
}
\examples{
observed <- c(1,5,6,8,9,20)
predicted <- c(2,5,6,7,9,22)
goof(observed,predicted)
goof(observed,predicted,coefficient = c('R2','RMSE','RPD'),cex=3,pch=19,col='blue')
}
\author{
Brendan Malone
}

