% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compSpec.R
\name{compSpec}
\alias{compSpec}
\title{Reduce column dimensions}
\usage{
compSpec(spectra, window)
}
\arguments{
\item{spectra}{matrix where each row is a spectrum and each column a wavelength}

\item{window}{window size over which the spectra will be averaged}
}
\description{
This function reduces the dimensions of a eSpectra by averaging values inside a window of a specified size
}
\examples{
#load data
data("soilSpecDat")

#subset spectra
spectra <- soilSpecDat[,grepl('[X]',colnames(soilSpecDat))]
colnames(spectra) <- substr(colnames(spectra),2,5)
wavelengths <- as.numeric(colnames(spectra))

#plot spectrum
plot(wavelengths,spectra[1,],type='b',ylab= "reflectance")

spectra_2 <- compSpec(spectra,9)
wavelength_2 <- colnames(spectra_2)

#plot processed spectrum
points(wavelength_2,spectra_2[1,],cex=2,col='red',pch=21,bg='blue') 


}
\author{
Brendan Malone
}
