% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_SoilSpectra.R
\name{filter_SoilSpectra}
\alias{filter_SoilSpectra}
\title{Filtering methods for SoilSpectra objects}
\usage{
filter_SoilSpectra(SoilSpectra, type, n = 11, p = 2, m = 0, res = NULL,
  ...)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}

\item{type}{\itemize{
\item 'S-Golay' for Sativsky Golay filter using \code{\link{sgolayfilt}},
\item 'Wavelet' for Wavelet decomposition using \code{\link{accessC.wd}},
\item 'MSC' for Multiplicative scatter correction, 
\item 'SNV' for Standard Normal Variate transformation.
\item 'C-hull' for Continuum removal by removing the convex hull
}}

\item{If}{type = 'S-Golay' then n,p,m parameters control \code{\link{sgolayfilt}} window size, polynomial order and derivative order respectively.}

\item{If}{type = 'Wavelet' res level to be extracted from wavelet decomposition model, see \code{\link{accessC.wd}}}

\item{If}{type = 'C-hull' specType also needs to be specified, with 0 if the data is in absorbance units or 1 if the data is in reflectance units.}

\item{If}{type is a character vector with more than one filter type the method will apply all the filters in the order they are specified.}
}
\description{
This method calls different filters on matrices in which each row is a separate spectrum
depending on the value of 'type':
}
\examples{
\dontrun{
data("SoilSpectraExample")


plot(SoilSpectraExample@Wavelength,
     SoilSpectraExample@Spectra[1,],
     main=SoilSpectraExample@Treatments,
     ylab=SoilSpectraExample@Type,
     xlab=slotNames(SoilSpectraExample)[3])


#Savitsky-Golay filter
SGolay_Spectra <- filter_SoilSpectra(SoilSpectraExample,type = 'S-Golay')

plot(SGolay_Spectra@Wavelength,
     SGolay_Spectra@Spectra[1,],
     main=SGolay_Spectra@Treatments,
     ylab=SGolay_Spectra@Type,
     xlab=slotNames(SGolay_Spectra)[3])

#Multiplicative scatter correction

MSC_Spectra <- filter_SoilSpectra(SoilSpectraExample,type = 'MSC')

plot(MSC_Spectra@Wavelength,
     MSC_Spectra@Spectra[1,],
     main=MSC_Spectra@Treatments,
     ylab=MSC_Spectra@Type,
     xlab=slotNames(MSC_Spectra)[3])



#Standard Normal Variate transform

SNV_Spectra <- filter_SoilSpectra(SoilSpectraExample,type = 'SNV')

plot(SNV_Spectra@Wavelength,
     SNV_Spectra@Spectra[1,],
     main=SNV_Spectra@Treatments,
     ylab=SNV_Spectra@Type,
     xlab=slotNames(SNV_Spectra)[3])


#Convex-hull transform

CHull_Spectra <- filter_SoilSpectra(SoilSpectraExample,type = 'C-hull')

plot(CHull_Spectra@Wavelength,
     CHull_Spectra@Spectra[1,],
     main=CHull_Spectra@Treatments,
     ylab=CHull_Spectra@Type,
     xlab=slotNames(CHull_Spectra)[3])
}
}
\references{
Savitzky, Abraham and Golay, M. J. E. (1964) Smoothing and Differentiation of Data by Simplified Least Squares Procedures, Analytical Chemistry, 36, pp. 1627-1629.
}
\author{
Brendan Malone, Budiman Minasny, Michael Nelson, Sebastian Campbell and Mario Fajardo.
}
