% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mscBLC.R
\name{mscBLC}
\alias{mscBLC}
\title{Multiplicative scatter correction (MSC)
 NOTE: This function will be deprecated in the next version, please use \code{\link{filter_spectra}} instead
 @param spectra a matrix or data.frame with wavelengths as columns and spectra as rows
 @export}
\usage{
mscBLC(spectra)
}
\description{
Multiplicative scatter correction (MSC)
 NOTE: This function will be deprecated in the next version, please use \code{\link{filter_spectra}} instead
 @param spectra a matrix or data.frame with wavelengths as columns and spectra as rows
 @export
}

