% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,SoilSpectra,ANY-method}
\alias{plot,SoilSpectra,ANY-method}
\title{plot method for SoilSpectra objects}
\usage{
\S4method{plot}{SoilSpectra,ANY}(x, y, ...)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}

\item{ID}{numeric or character vector with ID of spectra to plot (default=1).}
}
\description{
This method generates a plot.SoilSpectra for a Spectra Object
}
\examples{
 
\dontrun{
data("SoilSpectraExample")
plot(SoilSpectraExample)
plot(SoilSpectraExample,type='l')
}
}
\author{
Mario Fajardo.
}
