% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\docType{methods}
\name{plot,eSpectra-method}
\alias{plot,eSpectra-method}
\title{plot method for eSpectra objects}
\usage{
\S4method{plot}{eSpectra}(x, y, ...)
}
\arguments{
\item{eSpectra}{Object of class \code{\link{eSpectra}}.}

\item{ID}{numeric or character vector with ID of spectra to plot (default=1).}
}
\description{
This method generates a plot.eSpectra for a Spectra Object
}
\examples{
 
\dontrun{
data("eSpectraExample")
plot(eSpectraExample)
plot(eSpectraExample,type='l')
}
}
\author{
Mario Fajardo.
}
