% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\docType{methods}
\name{points,SoilSpectra-method}
\alias{points,SoilSpectra-method}
\title{points method for SoilSpectra objects}
\usage{
\S4method{points}{SoilSpectra}(x, ...)
}
\arguments{
\item{SoilSpectra}{Object of class \code{\link{SoilSpectra}}.}

\item{ID}{numeric or character vector with ID of spectra to plot (default=1).}
}
\description{
This method generates a plot.SoilSpectra for a Spectra Object
}
\examples{
 
\dontrun{
data("SoilSpectraExample")
plot(SoilSpectraExample)
points(SoilSpectraExample)
}
}
\author{
Mario Fajardo.
}
