% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\docType{methods}
\name{show,eSpectra-method}
\alias{show,eSpectra-method}
\title{show method for eSpectra objects}
\usage{
\S4method{show}{eSpectra}(object)
}
\arguments{
\item{eSpectra}{Object of class \code{\link{eSpectra}}.}
}
\description{
This method generates a description for a eSpectra Object
}
\author{
Mario Fajardo.
}
