% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvBLC.R
\name{snvBLC}
\alias{snvBLC}
\title{Standard Normal Variate (SNV) transformation}
\usage{
snvBLC(spectra)
}
\arguments{
\item{spectra}{a matrix or data.frame with wavelengths as columns and spectra as rows}
}
\description{
Standard Normal Variate (SNV) transformation
}
\note{
This function will be deprecated in the next version, please use \code{\link{filter_spectra}} instead
}

