% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra2colour.R
\name{spectra2colour}
\alias{spectra2colour}
\title{Method that extracts RGB and Munsell colour from \code{\link{SoilSpectra}} objects.}
\usage{
spectra2colour(SoilSpectra, ...)
}
\arguments{
\item{SoilSpectra}{A \code{\link{SoilSpectra}} object.}
}
\description{
Converts spectra reflectance into RGB and Munsell colours
}
\examples{
\dontrun{
data("SoilSpectraExample")

spectra2colour(SoilSpectraExample)

spectra2colour(SoilSpectraExample[23]) 
}
}
\author{
Michael Nelson and Mario Fajardo
}
